Reworked connection generation

+++ apphub-lib/apphub-lib-geometry/src/main/java/com/github/saphyra/apphub/lib/geometry/Coordinate.java
@@ -3,11 +3,9 @@ package com.github.saphyra.apphub.lib.geometry;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 @Data
 public class Coordinate {
-    private final BigDecimal x;
-    private final BigDecimal y;
+    private final double x;
+    private final double y;
 }
